requirements.txt
streamlit==1.27.2
pandas==2.1.1
plotly==5.17.0
numpy==1.26.1
Pillow==10.1.0
openpyxl==3.1.2
statsmodels==0.14.0
scikit-learn==1.3.1
st-pages==0.4.5